requirements.txt
numpy == 1.24.3
tensorflow-intel == 2.13.0
pandas == 2.0.3
keras == 2.13.1
streamlit == 1.24.1
pickle-mixin == 1.0.2
scikit-learn == 1.3.0
click>=7.0,<8.0